Доработка классов Mysqlobjects
